MAINT: move bidsmodel and bidsview to library layer
